+++ src/main/kotlin/me/rafaelldi/aspire/services/AspireHostServiceViewDescriptor.kt
@@ -22,6 +22,10 @@ class AspireHostServiceViewDescriptor(
 ) : ServiceViewDescriptor, DataProvider {
 
     private val toolbarActions = DefaultActionGroup(
+        ActionManager.getInstance().getAction("Aspire.Host.Run"),
+        ActionManager.getInstance().getAction("Aspire.Host.Debug"),
+        ActionManager.getInstance().getAction("Aspire.Host.Stop"),
+        Separator(),
         ActionManager.getInstance().getAction("Aspire.Manifest"),
         ActionManager.getInstance().getAction("Aspire.Dashboard"),
         ActionManager.getInstance().getAction("Aspire.Diagram"),