+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - [#177](https://github.com/rafaelldi/aspire-plugin/issues/177): Support `commandLineArgs` for the host projects from `launchSettings.json`
+- New actions to the dashboard to run, debug and stop the host project
 
 ## [1.0.1] - 2024-05-27